Update README.md with comprehensive project overview
This commit enhances the README.md file by providing a detailed overview of the Predict Health Costs with Regression project. It includes sections on the dataset used, methodology encompassing data preparation, exploratory data analysis, and model building, as well as the results achieved after training the model.

The Usage section has been updated with clear instructions on how to install dependencies and import the necessary Python libraries, facilitating ease of replication for users interested in running the project on their own systems.

This update aims to offer a clear and concise guide to understanding the project's objectives, the approach taken to achieve these objectives, and the outcomes of the modeling process. Additionally, it seeks to assist users in navigating the project's codebase by providing step-by-step instructions on setting up and executing the project.